Split model training to several modules
